+++ mrbgems/picoruby-net/mrblib/ntp.rb
@@ -25,12 +25,12 @@ class Net
       # Create a NTP packet
       ntp_packet = String.pack('C', 0x1b) + ("\0" * 47)
 
-      response = UDPClient.send(ntp_server, ntp_port, ntp_packet, false)
+      r = UDPClient.send(ntp_server, ntp_port, ntp_packet, false)
 
-      if response && response.size == NTP_PACKET_SIZE
+      if r && r.size == NTP_PACKET_SIZE
         # Extract NTP timestamp (seconds from 1900-01-01)
-        seconds = ((response[40]&.ord||0) << 24) | ((response[41]&.ord||0) << 16) | ((response[42]&.ord||0) << 8) | (response[43]&.ord||0)
-        fraction = ((response[44]&.ord||0) << 24) | ((response[45]&.ord||0) << 16) | ((response[46]&.ord||0) << 8) | (response[47]&.ord||0)
+        seconds = ((r[40]&.ord||0) << 24) | ((r[41]&.ord||0) << 16) | ((r[42]&.ord||0) << 8) | (r[43]&.ord||0)
+        fraction = ((r[44]&.ord||0) << 24) | ((r[45]&.ord||0) << 16) | ((r[46]&.ord||0) << 8) | (r[47]&.ord||0)
 
         # Convert NTP timestamp to UNIX timestamp
         unixtime = (seconds - 2_208_988_800 + (fraction / 2.0**32)).to_i
