+++ mrbgems/picoruby-net/ports/rp2040/include/common.h
@@ -1,0 +1,14 @@
+#ifndef NET_COMMON_DEFINED_H_
+#define NET_COMMON_DEFINED_H_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+err_t Net_get_ip(const char *name, ip_addr_t *ip);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // NET_COMMON_DEFINED_H_
