--- mrbgems/picoruby-net/ports/rp2040/net.c
@@ -1,13 +1,0 @@
-#include "pico/cyw43_arch.h"
-
-void
-lwip_begin(void)
-{
-  cyw43_arch_lwip_begin();
-}
-
-void
-lwip_end(void)
-{
-  cyw43_arch_lwip_end();
-}
