+++ mrbgems/picoruby-net/src/net.c
@@ -1,58 +1,55 @@
+#include <stdbool.h>
+#include "mruby.h"
+#include "mruby/presym.h"
+#include "mruby.h"
 #include "../include/net.h"
 
-
-static void
-dns_found(const char *name, const ip_addr_t *ip, void *arg)
+static mrb_value
+mrb_net_dns_s_resolve(mrb_state *mrb, mrb_value self)
 {
-  ip_addr_t *result = (ip_addr_t *)arg;
-  if (ip) {
-    ip4_addr_copy(*result, *ip);
-  } else {
-    ip4_addr_set_loopback(result);
-  }
+  const char *host;
+  mrb_bool is_tls;
+  mrb_get_args(mrb, "zb", &host, &is_tls);
+  return DNS_resolve(mrb, host, is_tls);
 }
 
-static err_t
-get_ip_impl(const char *name, ip_addr_t *ip)
+static mrb_value
+mrb_net_tcpclient_s__request_impl(mrb_state *mrb, mrb_value self)
 {
-  lwip_begin();
-  err_t err = dns_gethostbyname(name, ip, dns_found, ip);
-  lwip_end();
-  return err;
+  const char *host;
+  mrb_int port;
+  mrb_value data;
+  mrb_bool use_dtls;
+  mrb_get_args(mrb, "ziob", &host, &port, &data, &use_dtls);
+  return TCPClient_send(host, port, mrb, data, use_dtls);
 }
 
-err_t
-Net_get_ip(const char *name, ip_addr_t *ip)
+static mrb_value
+mrb_net_udpclient_s__send_impl(mrb_state *mrb, mrb_value self)
 {
-  err_t err = get_ip_impl(name, ip);
-  if (err != ERR_OK) {
-    return err;
-  }
-  while (!ip_addr_get_ip4_u32(ip)) {
-    sleep_ms(50);
-  }
-  return ERR_OK;
+  const char *host;
+  mrb_int port;
+  mrb_value data;
+  mrb_bool use_dtls;
+  mrb_get_args(mrb, "ziob", &host, &port, &data, &use_dtls);
+  return UDPClient_send(host, port, mrb, data, use_dtls);
 }
 
-static void
-DNS_resolve(const char *name, char* ipaddr, bool is_tcp)
+void
+mrb_picoruby_net_gem_init(mrb_state* mrb)
 {
-  (void)is_tcp;
-  ip_addr_t ip;
-  ip4_addr_set_zero(&ip);
-  Net_get_ip(name, &ip);
-  if(!ip4_addr_isloopback(&ip)) {
-    ipaddr_ntoa_r(&ip, ipaddr, 16);
-  }
+  struct RClass *class_Net = mrb_define_class_id(mrb, MRB_SYM(Net), mrb->object_class);
+
+  struct RClass *class_Net_DNS = mrb_define_class_under_id(mrb, class_Net, MRB_SYM(DNS), mrb->object_class);
+  mrb_define_class_method_id(mrb, class_Net_DNS, MRB_SYM(resolve), mrb_net_dns_s_resolve, MRB_ARGS_REQ(2));
+
+  struct RClass *class_Net_TCPClient = mrb_define_class_under_id(mrb, class_Net, MRB_SYM(TCPClient), mrb->object_class);
+  mrb_define_class_method_id(mrb, class_Net_TCPClient, MRB_SYM(_request_impl), mrb_net_tcpclient_s__request_impl, MRB_ARGS_REQ(4));
+
+  struct RClass *class_Net_UDPClient = mrb_define_class_under_id(mrb, class_Net, MRB_SYM(UDPClient), mrb->object_class);
+  mrb_define_class_method_id(mrb, class_Net_UDPClient, MRB_SYM(_send_impl), mrb_net_udpclient_s__send_impl, MRB_ARGS_REQ(4));
 }
-
-#if defined(PICORB_VM_MRUBY)
-
-#include "mruby/net.c"
-
-#elif defined(PICORB_VM_MRUBYC)
-
-#include "mrubyc/net.c"
-
-#endif
-
+void
+mrb_picoruby_net_gem_final(mrb_state* mrb)
+{
+}
